requirements.txt
numpy
pandas
opencv-python
pytesseract
openai
anthropic
pillow
requests
nltk
weighted-levenshtein
google-generativeai
python-dotenv